Day 131 Mean Median and Mode using Python
Mean, Median and Mode are the fundamentals of statistics used in almost every domain where we deal with numbers. Python is one of the best programming languages for numerical calculations. So you should know how to calculate mean, median and mode using Python without using any built-in Python library or module.  I will take you through how to calculate mean, median, and mode using Python.


Mean Median and Mode using Python
Mean
The mean is the average value of all the values in a dataset. To calculate the mean value of a dataset, we first need to find the sum of all the values and then divide the sum of all the values by the total number of values. So here’s how to calculate the mean using Python:

Median
The Median is the middle value among all the values in sorted order. Here we need to calculate the mid-value of all the values in a dataset. But before calculating the Median, we need to arrange all the values in sorted order. There are two different ways of calculating the median value:

when the total number of values is even: Median  = [(n/2)th term + {(n/2)+1}th]/2
when the total number of values is odd: Median = {(n+1)/2}thterm
Now below is how you can calculate the median using Python:


Mode
Mode is the most frequently occurring value among all the values. Below is how we can calculate the mode value of a dataset using Python:

Mode
Mode is the most frequently occurring value among all the values. Below is how we can calculate the mode value of a dataset using Python: