Day 132 Execution Time of a Python Program
The execution or running time of the program indicates how quickly the output is delivered based on the algorithm you used to solve the problem. To calculate the execution time of the program, we need to calculate the time taken by the program from its initiation to the final result. So if you want to learn how to calculate the execution time of a Python program, this article is for you. I will take you through a tutorial on calculating the execution time of a Python program.

Execution Time of a Python Program
It is important to calculate the execution time when working on a large project. When working on a large project, we have several approaches in mind. The best should be the one that takes the shortest execution time in all scenarios.

So to calculate the execution time of a Python program, we need to follow the steps mentioned below:

1. First, store the time of initiation of the program into a variable;
2. Write the Python program;
3. Store the end time of the program into a variable;
4. Subtract the time of initiation of the program from the end time of the program;

In the end, you will get the execution time of your program in seconds.

Calculating Execution Time of a Python Program
Now let’s follow the process described in the above section to calculate the time taken by a Python program. Here I will write a simple program to create acronyms: