Day 133 Count Number of Words in a Column using Python
While working on a Data Science task, sometimes we have to deal with textual data. One of the problems beginners face while working on a textual dataset is counting the number of words in a piece of text. So if you want to learn how to count the number of words in a textual dataset, this article is for you.  I will take you through a tutorial on how to count the number of words in a column using Python.


Count Number of Words in a Column using Python

Most data science professionals use the pandas library for data handling and preparation. The pandas library doesn’t have any method to count the number of words in a piece of text. One way to solve this problem is by finding the length of the text by splitting the complete text.

So let’s import a textual dataset where we can count the number of words in a column:

So the dataset has three columns now, Article, Title and the Number of Words that contains the number of words in the article column.

Summary
The pandas library doesn’t have any method to count the number of words in a piece of text. One way to solve this problem is by finding the length of the text by splitting the complete text. So, this is how you can count the number of words in any column while working on a textual dataset. I hope you liked this article on counting the number of words in a column using Python. Feel free to ask valuable questions in the comments section below.