Day 136 Visualize Linear Relationship using Python
A linear relationship is a statistical term that is nothing but the relationship between two variables. A linear relationship shows how well two variables x and y are related to each other. As a data science professional, you should know how to visualize a linear relationship as it will show the relationship between two numerical features of a dataset. So if you want to learn how to visualize a linear relationship, this article is for you. In this article, I will take you through a tutorial on how to visualize a linear relationship using Python.

Visualize a Linear Relationship using Python
When the value of variable increases or decreases with the increase or decrease in the value of another variable, then it is nothing but a linear relationship. When we visualize a linear relationship, it shows whether the relationship between the two features is linear or not.

You can use any data visualization library in Python to visualize a linear relationship. I prefer to use plotly as it provides interactive results. But as so many Python programmers use matplotlib for data visualization, I will show you how to visualize a linear relationship with Python using plotly and matplotlib

Visualizing Linear Relationships using Python
So let’s import a dataset and all the necessary Python libraries for this task:

Here’s how to visualize linear relationships by using the plotly library in Python:

To visualize linear relationships using matplotlib, you have to use seaborn.regplot method. So here’s how to plot linear relationships by using the matplotlib library in Python:


Summary
So this is how to visualize linear relationships using the Python programming language. When the value of variable increases or decreases with the increase or decrease in the value of another variable, then it is nothing but a linear relationship. I hope you liked this article on visualizing linear relationships using Python. Feel free to ask valuable questions in the comments section below.