Day 139 Reverse a String using Python.
A string is a sequence of characters enclosed in single or double-quotes. String inversion is one of the most common problems in computer science. Here we need to reverse the characters of a string. So, if you want to learn how to reverse a string, this article is for you. In this article, I will walk you through a tutorial on how to reverse a string using Python.

Reverse a String using Python
There are many ways to reverse a string using Python. You can use any method that you find easy unless you are told to use a specific method.

You must have heard of the concept of slicing in Python. Here I will show you how to use string slicing to reverse a string using Python:

The first character in the string has index 0, and the last character has index n-1, where n is the length of the string. The string slicing operator “::” reads all the characters of the string, and -1, in the end, reverses the order of the characters. This is how we can reverse a string.

Summary
So this is how we can use string slicing for reversing the order of the characters of a string. String inversion is one of the most common problems in computer science. I hope you liked this article on a tutorial on reversing a string using the Python programming language. Feel free to ask valuable questions in the comments section below.