Day 146 Age Calculator using Python
Age Calculator is an amazing coding project idea for beginners. If you are new to any programming language, you should try making an age calculator. It is an application where a user enters his date of birth as an input, and the application gives his age as an output. So, if you want to learn how to make an age calculator using the Python programming language, this article is for you. I will introduce you to a tutorial on how to create an age calculator using Python.

Age Calculator using Python
Age Calculator is an amazing application to create as a beginner in any programming language. To create an age calculator, you need two dates:

today’s date
date of birth
You can either ask the user for both dates or just ask for the date of birth and use today’s date from the computer itself. Asking for the birthday only seems like a more user-friendly option. So here’s how to create an age calculator using Python:

In the above code:

I have first defined a Python function where I am asking for three user inputs:
y: year of birth 
m: month of birth 
d: date of birth
Then I am importing the datetime module in Python inside the function
Then in the next line, I am taking today’s date by using the datetime.now() method of the datetime module
Then I have introduced a new variable in the next line as dob, where I am using the date of birth as the input given by the user
Then I am subtracting the dob with today’s date and then dividing it by 365.25 which is returning the age of the user.

Summary
Age Calculator is an amazing coding project idea for beginners. It is an application where a user enters his date of birth as an input, and the application gives his age as an output. I hope you liked this article on how to create an age calculator using Python. Feel free to ask your valuable questions in the comments section below.