Add files via upload
Price is one of the most important factors influencing the demand for a product. Elasticity refers to the degree of response, and the price elasticity of demand refers to the degree of responsiveness of demand for a product due to the change in its price. We study the price elasticity of demand in economics, but if you want to learn how to calculate the price elasticity of demand using Python, I will give you a tutorial on calculating the price elasticity of demand using Python.

Price Elasticity of Demand
Price elasticity of demand refers to the degree of responsiveness of demand for a product to a change in price. Simply put, it means the degree to which the demand for a product changes with an increase or decrease in its price. For example, the demand for a product increases by 20% due to a 10% decrease in its price. This is what it means to change in demand with the change in the price of a product. And when you calculate the degree to which demand changes, it’s called the price elasticity of demand.

To calculate the price elasticity of demand, you have to use the formula mentioned below:

Percentage Change in Quantity Demanded / Percentage Change in the Price

I hope you now have understood what price elasticity means and how to calculate it. Now in the section below, I will take you through a tutorial on how to calculate, price elasticity of demand using Python.

Price Elasticity of Demand using Python
I will start the task of calculating price elasticity of demand using Python by creating a small dataset that should contain data about the change in the price and the demand for a product:

The first rows of this dataset contain the initial demand and price for a product, and the subsequent rows contain the change in demand and the change in the price of the product. Now our next step is to add two more columns as the Percentage change in Demand and Percentage change in Price by calculating them:

Now the last step is to calculate the price elasticity of demand (% Change in Demand / % Change in Price) by adding a new column to this data. Below is how you can calculate it using Python:

Summary
Price elasticity of demand refers to the degree of responsiveness of demand for a product to a change in price. Simply put, it means the degree to which the demand for a product changes with an increase or decrease in its price. I hope you liked this article on how to calculate the price elasticity of demand using Python. Feel free to ask your valuable questions in the comments section below.